FAC-210: Service improved to show suitable error, when user is not in a member in the group

+++ flickit-assessment-users/src/main/java/org/flickit/assessment/users/application/service/expertgroupaccess/DeleteExpertGroupMemberService.java
@@ -2,15 +2,18 @@ package org.flickit.assessment.users.application.service.expertgroupaccess;
 
 import lombok.AllArgsConstructor;
 import org.flickit.assessment.common.exception.AccessDeniedException;
+import org.flickit.assessment.common.exception.ResourceNotFoundException;
 import org.flickit.assessment.users.application.port.in.expertgroupaccess.DeleteExpertGroupMemberUseCase;
 import org.flickit.assessment.users.application.port.out.expertgroup.LoadExpertGroupOwnerPort;
 import org.flickit.assessment.users.application.port.out.expertgroupaccess.DeleteExpertGroupMemberPort;
+import org.flickit.assessment.users.application.port.out.expertgroupaccess.LoadExpertGroupAccessPort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Objects;
 import java.util.UUID;
 
+import static org.flickit.assessment.users.common.ErrorMessageKey.DELETE_EXPERT_GROUP_MEMBER_USER_ID_NOT_FOUND;
 import static org.flickit.assessment.common.error.ErrorMessageKey.COMMON_CURRENT_USER_NOT_ALLOWED;
 
 @Service
@@ -19,11 +22,15 @@ import static org.flickit.assessment.common.error.ErrorMessageKey.COMMON_CURRENT
 public class DeleteExpertGroupMemberService implements DeleteExpertGroupMemberUseCase {
 
     private final LoadExpertGroupOwnerPort loadExpertGroupOwnerPort;
+    private final LoadExpertGroupAccessPort loadExpertGroupAccessPort;
     private final DeleteExpertGroupMemberPort deleteExpertGroupMemberPort;
 
     @Override
     public void deleteMember(Param param) {
         validateCurrentUser(param.getExpertGroupId(), param.getCurrentUserId());
+        var access = loadExpertGroupAccessPort.loadExpertGroupAccess(param.getExpertGroupId(), param.getUserId());
+        if(access.isEmpty())
+            throw new ResourceNotFoundException(DELETE_EXPERT_GROUP_MEMBER_USER_ID_NOT_FOUND);
         deleteExpertGroupMemberPort.deleteMember(param.getExpertGroupId(), param.getUserId());
     }
